// src/lib/graphql/generated/graphql.ts
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** An ISO 8601-encoded date */
  ISO8601Date: { input: any; output: any; }
  /** An ISO 8601-encoded datetime */
  ISO8601DateTime: { input: any; output: any; }
  Upload: { input: any; output: any; }
  /** A valid URL, transported as a string */
  Url: { input: any; output: any; }
};
// ...
export type Duration = {
  __typename?: 'Duration';
  minuteUnitName?: Maybe<Scalars['String']['output']>;
  minutes?: Maybe<Scalars['Int']['output']>;
  secondUnitName?: Maybe<Scalars['String']['output']>;
  seconds?: Maybe<Scalars['Int']['output']>;
};
// ...
export type OriginalVideo = Video & {
  __typename?: 'OriginalVideo';
  categories?: Maybe<Array<Category>>;
  comment?: Maybe<Comment>;
  commentaryUrl?: Maybe<Scalars['String']['output']>;
  comments?: Maybe<CommentConnection>;
  commentsCount: Scalars['Int']['output'];
  contentEnd?: Maybe<Duration>;
  cookies?: Maybe<Cookies>;
  createdAt: Scalars['ISO8601DateTime']['output'];
  creator?: Maybe<Creator>;
  creators?: Maybe<Array<Creator>>;
  currentUserHasLike?: Maybe<Scalars['Boolean']['output']>;
  currentUserVideoRating?: Maybe<VideoRating>;
  description?: Maybe<Scalars['String']['output']>;
  drmDashFileUrl?: Maybe<Scalars['String']['output']>;
  drmHlsFileUrl?: Maybe<Scalars['String']['output']>;
  duration: Duration;
  fileUrl?: Maybe<Scalars['String']['output']>;
  filmarksUrl?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  jiqooNextVideo?: Maybe<OriginalVideo>;
  landscapeThumbnail?: Maybe<Scalars['String']['output']>;
  likeNum: Scalars['Int']['output'];
  nextVideo?: Maybe<OriginalVideo>;
  playlists?: Maybe<Array<Playlist>>;
  playlistsVideos?: Maybe<Array<PlaylistsVideo>>;
  portraitThumbnail?: Maybe<Scalars['String']['output']>;
  publishedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  satisfiedNum: Scalars['Int']['output'];
  status: VideoStatusEnum;
  subtitles?: Maybe<Array<SrtSubtitle>>;
  thumbnailCategory?: Maybe<Category>;
  title?: Maybe<Scalars['String']['output']>;
  trailer?: Maybe<TrailerVideo>;
  translations: Array<VideoTranslation>;
  updatedAt: Scalars['ISO8601DateTime']['output'];
  viewingTime?: Maybe<VideoViewingTime>;
  vtts?: Maybe<Array<VttSubtitle>>;
  willBePublishedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  willBeUnpublishedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
};
// ...
export type Query = {
  __typename?: 'Query';
  adminComments: CommentConnection;
  adminCoupons: CouponConnection;
  adminImages: AdminImageConnection;
  adminOpsConfig: OpsConfig;
  adminOriginalVideos: OriginalVideoConnection;
  adminScreenPopup?: Maybe<ScreenPopup>;
  adminScreenPopups: ScreenPopupConnection;
  adminTestMovies: Array<TestMovie>;
  adminUsers: UserConnection;
  adminWithdrawalVideos: Array<WithdrawalVideo>;
  allOriginalVideos: Array<OriginalVideo>;
  category: Category;
  comment?: Maybe<Comment>;
  commentReplies: CommentRepliesConnection;
  coupon?: Maybe<Coupon>;
  currentUser?: Maybe<User>;
  diveruOriginalVideos: OriginalVideoConnection;
  firstRecommendationRow: RecommendationRow;
  homeScreens: Array<HomeScreen>;
  invitationCounts: InvitationCount;
  jiqooCategory: Category;
  jiqooCommentReplies: CommentRepliesConnection;
  jiqooHomeScreens: Array<HomeScreen>;
  jiqooOriginalVideo?: Maybe<OriginalVideo>;
  jiqooOriginalVideos: OriginalVideoConnection;
  jiqooRootCategories: Array<RootCategory>;
  jiqooScreenPopups: ScreenPopupConnection;
  jiqooTrailerVideo?: Maybe<TrailerVideo>;
  jiqooTrailerVideos: TrailerVideoConnection;
  jiqooVideoComments: VideoCommentsConnection;
  jiqooVideoRecommends?: Maybe<VideoRecommendsConnection>;
  notifications: UserNotificationConnection;
  onboardingRecommendVideos?: Maybe<Array<OriginalVideo>>;
  originalVideo?: Maybe<OriginalVideo>;
  originalVideos: OriginalVideoConnection;
  payment: Payment;
  playlist: Playlist;
  playlists: Array<Playlist>;
  promotionCode?: Maybe<PromotionCode>;
  recommendationRow: RecommendationRow;
  recommendationRowList: RecommendationRowList;
  recommendedSearchKeywords: Array<RecommendedSearchKeyword>;
  reply: Reply;
  rootCategories: Array<RootCategory>;
  screenPopups: ScreenPopupConnection;
  trailerVideo?: Maybe<TrailerVideo>;
  trailerVideos: TrailerVideoConnection;
  trialVideos: Array<OriginalVideo>;
  tvSearchCategories: Array<TvSearchCategory>;
  userByAtId?: Maybe<User>;
  userInterestAnswers: Array<UserInterestAnswer>;
  userNotificationSettings: Array<UserNotificationSetting>;
  videoAiRecommends?: Maybe<VideoRecommendsConnection>;
  videoComments: VideoCommentsConnection;
  videoHistories: VideoHistoryConnection;
  videoRatings: VideoRatingConnection;
  videoRecommends?: Maybe<VideoRecommendsConnection>;
  withdrawalOriginalVideos: Array<OriginalVideo>;
  withdrawalPageVideos: Array<OriginalVideo>;
  withdrawalVideos: Array<TrailerVideo>;
};
// ...
export type GetOriginalVideoQueryVariables = Exact<{
  id: Scalars['ID']['input'];
}>;


export type GetOriginalVideoQuery = { __typename?: 'Query', originalVideo?: { __typename?: 'OriginalVideo', title?: string | null, description?: string | null, landscapeThumbnail?: string | null, likeNum: number, duration: { __typename?: 'Duration', minutes?: number | null, seconds?: number | null } } | null };
// ...
export const GetOriginalVideoDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"getOriginalVideo"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"id"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ID"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"originalVideo"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"id"},"value":{"kind":"Variable","name":{"kind":"Name","value":"id"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"landscapeThumbnail"}},{"kind":"Field","name":{"kind":"Name","value":"likeNum"}},{"kind":"Field","name":{"kind":"Name","value":"duration"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"minutes"}},{"kind":"Field","name":{"kind":"Name","value":"seconds"}}]}}]}}]}}]} as unknown as DocumentNode<GetOriginalVideoQuery, GetOriginalVideoQueryVariables>;